+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * fix: Keep using collection/collectionView user defined to access the interface within the Collection/AICollectionView object
 * feat: support chunk_num and section_num when uploading files in a collection
 * feat: support to check file byte size when uploading files in a collection or collectionView
+* feat: support to set shardNum and replicaNum for collectionView
 
 ## v1.7.0
 * feat: support /index/drop interface